Add more documentation and diagnostics
